Fix warning about different integer types in PN5180::readEEprom()

+++ PN5180.cpp
@@ -245,7 +245,7 @@ bool PN5180::readEEprom(uint8_t addr, uint8_t *buffer, int len) {
   PN5180DEBUG(len);
   PN5180DEBUG(F("...\n"));
 
-  uint8_t cmd[] = { PN5180_READ_EEPROM, addr, len };
+  uint8_t cmd[] = { PN5180_READ_EEPROM, addr, (uint8_t)len };
 
   SPI.beginTransaction(PN5180_SPI_SETTINGS);
   transceiveCommand(cmd, sizeof(cmd), buffer, len);
